Added sequence completion column
This adds a column to the Scheduler list, with light frame currently
captured, and required. These are connected to SchedulerJob attributes.

For each job, this also checks the number of light frames for each
sequence job in the schedule, in order to consolidate the total number
of frames effectively captured without dropping a sequence job by
mistake.