Reworking SchedulerJob cell widget management
Improved the shortcuts to update cell widgets in queueTable from
SchedulerJob instances, even if SchedulerJob should only be a data
model. Centralized updates in updateJobCell, but will somehow optimize
it by relocating the code to the relevant functions instead.

Also added a widget cell for the score, now displayed in queueTable.
Because negative scores are a bit of an ergonomy issue, those will
simply render as "<0". Basically, negative scores are a no-go for now.
Besides, displaying scores to the end-user now emphasizes on an issue
with job sorting, already hinted while refactoring predicates.