A bunch of FIXME, and cosmetics
Leaving the mount to track while the scheduler is sleeping could result in
problems if the end-user stops the scheduler or the system crashes. At the very
least the mount should not be requested to track.

Negative scores are weird for the end-user. Probably as soon as a negative
score is obtained, no further score evaluation should be attempted, just in
case a subsequent one adds enough to return positive.

Altitude cutoff has an issue with rising targets, which will get aborted.
Mitigated by a previous change which reschedules aborted jobs, but incorrect
nonetheless.

Manual execution of startup/shutdown scripts probably requires INDI to be
connected to succeed. Warning or notification should ensue properly.